# api_contracts/fastapi/openapi.yaml
openapi: 3.1.0
info:
  contact:
    email: juchiLee@proton.me
    name: Contracts for projects
    url: https://github.com/Dambarioid/University
  description: "My first OpenAPI contracts to link strictly front, back-end parts.\
    \ Also to give developers and users the whole vision what an endpoint does"
  license:
    name: GNU Affero General Public License v3.0 or later
    url: https://spdx.org/licenses/AGPL-3.0-or-later.html
  title: Contracts
  version: 1.0.1
servers:
- description: Frontend local server
  url: http://localhost:3000
- description: Backend local server
  url: http://localhost:3000
- description: Production server
  url: https://staging.gigantic-server.com/v1
paths:
  /auth/github:
    post:
      description: API for generating and sending GitHub token to sign via this one
      parameters:
      - description: Client Id of github
        explode: false
        in: path
        name: client_id
        required: true
        schema:
          type: string
        style: simple
      - description: Where user will redierected
        explode: false
        in: path
        name: redierect_uri
        required: true
        schema:
          type: string
        style: simple
      - description: What type of response will be got
        explode: false
        in: path
        name: response_type
        required: true
        schema:
          type: code
        style: simple
      - description: Which permissions will be granted to user
        explode: false
        in: path
        name: scope
        required: true
        schema:
          items:
            type: string
          type: array
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_auth_github_post_200_response"
          description: OK
  /register:
    post:
      description: API for creating new user and after logging it after this
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_register_post_200_response"
          description: OK
  /main/videos/categories:
    get:
      description: Getting from backend(S3) videos via recommendations
      parameters:
      - description: value must be `Bearer <token>` where `<token>` is api key prefixed
          with cal_
        explode: false
        in: header
        name: Authorization
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: OK
components:
  schemas:
    _register_post_200_response:
      example:
        GIT_TOKEN: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJuaWtpdGEiLCJleHAiOjE3NjI3NjkzNDB9.aBaKQVMaQ84BRwhEGLK3sUUOhtJGxm6t1R6cfO70IwE
      properties:
        GIT_TOKEN:
          description: generated with GitHub OAuth 2.0 helping
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJuaWtpdGEiLCJleHAiOjE3NjI3NjkzNDB9.aBaKQVMaQ84BRwhEGLK3sUUOhtJGxm6t1R6cfO70IwE
          title: GIT_TOKEN
          type: string
      required:
      - GIT_TOKEN
      title: _register_post_200_response
    _auth_github_post_200_response:
      example:
        GIT_TOKEN: https://github.com/login/oauth/authorize?client_id=%2A%2A%2A%2A%2A%2A%2A%2A%2A%2A&redirect_uri=http%3A%2F%2Flocalhost%3A3000&response_type=code&scope=%5B%27read%3Auser%27%2C%20%27user%3Aemail%27%5D
      properties:
        GIT_TOKEN:
          description: generated with GitHub OAuth 2.0 helping
          example: https://github.com/login/oauth/authorize?client_id=%2A%2A%2A%2A%2A%2A%2A%2A%2A%2A&redirect_uri=http%3A%2F%2Flocalhost%3A3000&response_type=code&scope=%5B%27read%3Auser%27%2C%20%27user%3Aemail%27%5D
          title: GIT_TOKEN
          type: string
      required:
      - GIT_TOKEN
      title: _auth_github_post_200_response